refactor: restructure pages with separate folders for each page
