+++ src/pages/settings/App.tsx
@@ -1,9 +1,9 @@
-function Autorzy() {
+function Ustawienia() {
   return (
     <div className="min-h-screen w-full px-8 text-center bg-gray-400 flex flex-col justify-center">
-      <h1 className="text-5xl font-bold">Autorzy</h1>
+      <h1 className="text-5xl font-bold">Ustawienia</h1>
     </div>
   );
 }
 
-export default Autorzy;
+export default Ustawienia;
